update page.tsx and layout.tsx + moved dashboard.tsx code to page.tsx
