+++ frontend/app/dashboard/layout.tsx
@@ -1,12 +1,8 @@
-import Dashboard from "@/app/ui/dashboard/dashboard"
 
 export default function Layout({children}: { children: React.ReactNode }) {
   return(
-      <div className="flex h-screen flex-col md:flex-row md:overflow-hidden">
-          <div className="w-full flex-none md:w-64">
-              <Dashboard />
-          </div>
-          <div className="flex-grow p-6 md:overflow-y-auto md:p-12">{children}</div>
+      <div className="flex h-screen md:flex-row">
+          {children}
       </div>
   );
 }
