+++ frontend/app/ui/dashboard/message.box.tsx
@@ -10,6 +10,11 @@ interface MessageBoxProps {
     isOpen: boolean
 }
 
+/**
+ * MessageBox component to display the messages and send new messages
+ * @param onClose - function to close the message box
+ * @param isOpen - boolean to check if the message box is open
+ */
 export function MessageBox({onClose, isOpen}: MessageBoxProps) {
     const [messages, setMessages] = useState<string[]>([]);
     const messagesEndRef = useRef<HTMLDivElement | null>(null);
@@ -41,18 +46,17 @@ export function MessageBox({onClose, isOpen}: MessageBoxProps) {
     }
 
     return (
-        <div className="bg-gray-200 w-full h-screen absolute left-0 top-0 p-6">
+        <div className="bg-gray-200 w-full h-screen p-6 md:w-64 flex-none md:relative absolute">
             <div className="flex justify-between md:hidden my-3">
                 <h2 className="text-2xl">Message History</h2>
                 <Button onClick={onClose}>X</Button>
             </div>
             <div className="overflow-auto h-4/6 mb-4 border border-gray-300 p-2">
-                {messages.map((message: string, index: number) => <div key={index}
-                                                                       className="bg-black text-sm text-gray-400 m-2 px-4 py-3 rounded relative"
-                                                                       role="alert">
-                    <strong className="font-bold">Message {index + 1}</strong><br/>
-                    <span className="block sm:inline">{message}</span>
-                </div>)}
+                {messages.map((message: string, index: number) =>
+                    <div key={index} className="bg-black text-sm text-gray-400 m-2 px-4 py-3 rounded">
+                        <strong className="font-bold">Message {index + 1}</strong><br/>
+                        <span className="block sm:inline">{message}</span>
+                    </div>)}
                 <div ref={messagesEndRef}/>
             </div>
             <form onSubmit={handleSubmit}>
@@ -60,11 +64,11 @@ export function MessageBox({onClose, isOpen}: MessageBoxProps) {
                     <Label htmlFor="message">Ask a filter question</Label>
                 </div>
                 <div className="flex justify-center m-2">
-                    <Textarea className="" id="message" name="message"
+                    <Textarea id="message" name="message"
                               placeholder={"Ask some more questions to filter your result"} required/>
                 </div>
                 <div className="flex justify-center m-2">
-                    <Button type="submit" className="">Search Routes</Button>
+                    <Button type="submit">Search Routes</Button>
                 </div>
             </form>
         </div>
