+++ frontend/app/ui/home/message.form.tsx
@@ -37,8 +37,8 @@ export default function MessageForm() {
             </div>
             <div className="flex justify-center m-2">
                 <Label htmlFor="message" className="text-sm text-center text-slate-400 w-8/12">Example:<br/>
-                    "Show me the routes from London to Paris on the 12th of December 2024 for 2 adults and 1 child.
-                    We plan to return between the 20th and 25th of December 2024."
+                    &quot; Show me the routes from London to Paris on the 12th of December 2024 for 2 adults and 1 child.
+                    We plan to return between the 20th and 25th of December 2024. &quot;
                 </Label>
             </div>
         </form>
